==> code/classification/pretraitements/03_personnages.ipynb <==
# Imports

import pandas as pd
import re

# Chargement des données

# Création d'un dossier appelé data
!mkdir data

# Téléchargement du fichier contenant les fanfictions et leurs métadonnées
!wget -P data https://gitlab.unistra.fr/paulinemoreau/memoire/-/raw/main/data/vrac/pretraitements/valide/fanfics_clean.csv

df_fanfic = pd.read_csv('data/fanfics_clean.csv')

df_fanfic.head()

# Création d'une colonne où les noms des personnages sont supprimés des body

Il est possible de récupérer les noms des personnages impliqués dans chaque fanfiction à l'aide de la colonne "character" créée lors du scraping.

En explorant rapidement le corpus collecté, j'ai remarqué que les auteurs de fanfictions peuvent utiliser différentes façons d'écrire les noms des personnages.
Par convention, ils écrivent les prénoms et noms de chaque personnages séparés par des virgules. Cependant, dans certains cas, quand un même personnage possède plusieurs dénominations différentes ils peuvent utiliser le séparateur "|" pour noter ces différentes dénominations. Aussi, quand un personnage possède un surnom, ce surnom est noté entre guillemets anglais. Enfin, les auteurs peuvent ajouter des détails ou informations sur le personnages qu'ils notent entre parenthèses.

Il est donc important de prendre en compte ces différents cas de figures pour récupérer correctement les différentes dénominations des personnages de chaque fanfiction.

###Fonction qui extrait les noms des personnages à l'intérieur d'une liste

# Fonction pour extraire les noms des personnages de chaque fanfiction
def clean_character_list(characters):
    if pd.isna(characters):
        return []

    # Remplacer les "|" par ","
    characters = characters.replace('|', ',')

    # Extraire les surnoms entre guillemets
    surnoms = re.findall(r'"(.*?)"', characters)

    # Supprimer les surnoms des noms principaux
    characters = re.sub(r'\s*"\s*.*?\s*"\s*', ' ', characters)

    # Supprimer les précisions entre parenthèses
    characters = re.sub(r'\s*\(.*?\)\s*', '', characters)

    # Transformer en liste propre
    character_list = [char.strip() for char in characters.split(',') if char.strip()]

    # Séparer prénom et nom pour chaque personnage
    character_list_sep_complet = []
    for char in character_list:
        # Ajouter nom et prénom séparément à la liste
        character_list_sep_complet.extend(char.split())

    # Ajouter les surnoms à la liste des noms à supprimer
    character_list_sep_complet.extend(surnoms)

    # Supprimer les doublons
    return list(set(character_list_sep_complet))

### Fonction qui supprime les noms de personnages d'un texte à partir d'une liste de ces noms

# Fonction pour supprimer les noms d'un texte
def remove_characters_from_text(texte, characters):
    if pd.isna(texte) or not characters:
        return texte

    for character in characters:
        # \b pour éviter de supprimer des bouts de mots
        texte = re.sub(r'\b' + re.escape(character) + r'\b', '', texte, flags=re.IGNORECASE)

    return texte

### Récupération des noms des personnages "nettoyés" dans la colonne "characters_clean"

Je rajoute une colonne contenant la liste des noms de personnages de chaque fanfiction au dataframe à l'aide de la fonction 'clean_character_list".

# Appliquer le nettoyage des noms dans une nouvelle colonne characters_clean
df_fanfic['characters_clean'] = df_fanfic['character'].apply(clean_character_list)

df_fanfic[['character', 'characters_clean']].head()

### Ajout d'une colonne body sans les noms des personnages (depuis body_clean)

J'utilise la nouvelle colonne 'characters_clean' et la fonction 'remove_characters_from_text' pour générer une nouvelle colonne 'body_no_chara' contenant le texte des fanfictions où les noms de personnages ont été supprimés.

# Supprimer les noms des fanfictions dans une nouvelle colonne body_no_characters
df_fanfic['body_no_chara'] = df_fanfic.apply(lambda row: remove_characters_from_text(row['body_clean'], row['characters_clean']), axis=1)

df_fanfic[['body_clean', 'body_no_chara']].head()

# Enregistrement
df_fanfic.to_csv('data/fanfics_perso.csv', index=False)
