==> code/classification/pretraitements/04_tokenisation.ipynb <==
# Imports

import pandas as pd
import spacy

!python -m spacy download fr_core_news_sm

# Chargement des données

# Création d'un dossier appelé data
!mkdir data

# Téléchargement du fichier contenant les fanfictions et leurs métadonnées
!wget -P data https://gitlab.unistra.fr/paulinemoreau/memoire/-/raw/main/data/vrac/pretraitements/valide/fanfics_perso.csv

df_fanfic = pd.read_csv('data/fanfics_perso.csv')

df_fanfic.head()

# Ajout de colonnes tokenisées

### Fonction de tokenisation

spacy_pipeline = spacy.load("fr_core_news_sm", disable=["parser", "ner", "lemmatizer"])

# Fonction de tokenisation sans mots vides
def split_into_tokens_wosw_spacy(text):
    doc = spacy_pipeline(text)
    return ' '.join([t.text for t in doc if not t.is_stop])

### Tokenisation du body nettoyé avec noms des personnages

Je rajoute une colonne 'tokenised_body' contenant les textes des fanfictions (avec noms des personnages) tokenisés en appliquant la fonction 'split_into_tokens_wosw_spacy' à la colonne 'body_clean".

df_fanfic['tokenised_body'] = df_fanfic['body_clean'].apply(split_into_tokens_wosw_spacy)

df_fanfic.head()

### Tokenisation du body nettoyé sans noms des personnages

Je fais de même pour les textes où les noms de personnages ont été supprimés pour obtenir les textes tokenisés dans la colonne 'tokenised_no_chara'.

df_fanfic['tokenised_no_chara'] = df_fanfic['body_no_chara'].apply(split_into_tokens_wosw_spacy)

df_fanfic.head()

# Enregistrement
df_fanfic.to_csv('data/fanfics_tokenised.csv', index=False)
