==> code/classification/pretraitements/05_lemmatisation.ipynb <==
# Imports


import pandas as pd
import spacy

!python -m spacy download fr_core_news_sm

# Pour mieux afficher les textes (colonnes plus larges)
pd.set_option('display.max_colwidth', 400)

# Chargement des données

# Création d'un dossier appelé data
!mkdir data
# Téléchargement du fichier fanfictions
!wget -P data https://gitlab.unistra.fr/paulinemoreau/memoire/-/raw/main/data/vrac/pretraitements/valide/fanfics_tokenised.csv

df_fanfic = pd.read_csv('data/fanfics_tokenised.csv')

df_fanfic.head()

# Lemmatisation

### Fonction de lemmatisation

spacy_pipeline = spacy.load("fr_core_news_sm")

# Fonction de lemmatisation avec spaCy
def split_into_lemmas(text):
    doc = spacy_pipeline(text)
    lemmatised_text = [w.lemma_ for w in doc]
    return ' '.join(lemmatised_text)

### Lemmatisation des body avec noms des personnages

Je rajoute une colonne 'lemmatised_body' contenant les textes des fanfictions (avec noms des personnages) lemmatisés en appliquant la fonction 'split_into_lemmas' à la colonne 'body_clean".

df_fanfic['lemmatised_body'] = df_fanfic['body_clean'].apply(split_into_lemmas)

df_fanfic.head()

### Lemmatisation des body sans noms des personnages

Je fais de même pour les textes où les noms de personnages ont été supprimés pour obtenir les textes lemmatisés dans la colonne 'lemmatised_no_chara'.

df_fanfic['lemmatised_no_chara'] = df_fanfic['body_no_chara'].apply(split_into_lemmas)

df_fanfic.head()

# Enregistrement
df_fanfic.to_csv('data/fanfics_lemmatised.csv', index=False)
