==> code/classification/pretraitements/06_limite_min.ipynb <==
# Imports

import pandas as pd

# Chargement des données



# Création d'un dossier appelé data
!mkdir data

# Téléchargement du fichier contenant les fanfictions et leurs métadonnées
!wget -P data https://gitlab.unistra.fr/paulinemoreau/memoire/-/raw/main/data/vrac/pretraitements/valide/fanfics_lemmatised.csv

df_fanfics = pd.read_csv('data/fanfics_lemmatised.csv')
df_fanfics.head()

len(df_fanfics)

# Analyse des fanfictions les plus courtes

# Calculer le nombre de fanfictions avec moins de 1000 caractères
court_fanfics = df_fanfics[df_fanfics['body_clean'].apply(len) < 1000]

# Afficher le nombre total de fanfictions qui dépassent la limite
total_court_fanfics = len(court_fanfics)
print(f"Nombre total de fanfictions avec moins de 1000 caractères : {total_court_fanfics}")

# Compter le nombre de fanfictions par tag
tag_counts = court_fanfics['tag'].value_counts()

# Afficher le nombre de fanfictions par tag
print("\nNombre de fanfictions par tag qui sont en dessous de la limite :")
print(tag_counts)

# Calculer le nombre de fanfictions avec moins de 500 caractères
court_fanfics = df_fanfics[df_fanfics['body_clean'].apply(len) < 500]

# Afficher le nombre total de fanfictions qui dépassent la limite
total_court_fanfics = len(court_fanfics)
print(f"Nombre total de fanfictions avec moins de 500 caractères : {total_court_fanfics}")

# Compter le nombre de fanfictions par tag
tag_counts = court_fanfics['tag'].value_counts()

# Afficher le nombre de fanfictions par tag
print("\nNombre de fanfictions par tag qui sont en dessous de la limite :")
print(tag_counts)

# Pour mieux afficher les textes
pd.set_option('display.max_colwidth', 500)

# Afficher le contenu des courtes fanfictions
print("\nContenu des fanfictions avec moins de 500 caractères :")
court_fanfics[['body']].head(12)

# Suppression des fanfictions trop courtes

Je choisis de supprimer les fanfictions dont le nombre de caractères est inférieur à 500. Je ne perds pas beaucoup de données, car elles sont peu nombreuses (12). De plus on peut voir ci-dessus que la plupart de ces fanfictions ne sont en fait pas vraiment des fanfictions, elles ne sont donc pas pertinentes et s'apparenteraient finalement plus à du bruit pour les algorithmes.

# Filtrer les fanfictions avec moins de 500 caractères
df_limit_min = df_fanfics[df_fanfics["nb_caracteres"] >= 500]

df_limit_min.head()

len(df_limit_min)

# Nombre de fanfictions par tags après les prétraitements

# Compter le nombre de fanfictions par tag
tag_counts = df_fanfics['tag'].value_counts()

# Afficher le résultat
print(f"Nombre de fanfictions par tag après prétraitements:\n")
print(tag_counts)

# Enregistrement
df_limit_min.to_csv('data/fanfics_min.csv', index=False)
